+++ scripts/etl/clear_state_inputs.py
@@ -10,6 +10,7 @@ from collections import deque
 import boto3
 from etl_utils.constants import CHANGELOG_NUMBER, WorkerKey
 from etl_utils.io import pkl_dumps_lz4
+from sds.epr.bulk_create.bulk_load_fanout import FANOUT
 
 from test_helpers.aws_session import aws_session
 from test_helpers.terraform import read_terraform_output
@@ -37,6 +38,12 @@ def main(changelog_number, workspace):
         s3_client.put_object(
             Bucket=etl_bucket, Key=WorkerKey.LOAD, Body=EMPTY_JSON_DATA
         )
+        for i in range(FANOUT):
+            s3_client.put_object(
+                Bucket=etl_bucket,
+                Key=f"{WorkerKey.LOAD}.{i}",
+                Body=pkl_dumps_lz4(EMPTY_JSON_DATA),
+            )
         s3_client.delete_object(Bucket=etl_bucket, Key=CHANGELOG_NUMBER)
 
     if changelog_number: