+++ src/etl/sds/tests/changelog/_test_changelog_scenarios.py
@@ -6,11 +6,11 @@ from domain.core.device import Device, DeviceTag
 from domain.repository.device_repository import DeviceRepository
 from domain.repository.keys import TableKey
 from domain.repository.marshall import unmarshall
+from etl.sds.worker.load_bulk.tests._test_load_bulk_worker import MockDeviceRepository
 from etl_utils.io import pkl_load_lz4
 from event.aws.client import dynamodb_client as get_dynamodb_client
 from mypy_boto3_s3 import S3Client
 
-from etl.sds.worker.load_bulk.tests._test_load_bulk_worker import MockDeviceRepository
 from test_helpers.terraform import read_terraform_output
 
 from .conftest import ETL_BUCKET, parametrize_over_scenarios